+++ Front-end/app/src/main/res/layout/activity_add_company.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".AddCompanyActivity">
+    tools:context=".ui.registration.AddCompanyActivity">
 
     <TextView
         android:id="@+id/addcompany_text_name"
