+++ Front-end/app/src/main/res/layout/activity_registration.xml
@@ -5,7 +5,7 @@
     android:id="@+id/registration_constraintlayout_constraint"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RegisterActivity">
+    tools:context=".ui.registration.RegisterActivity">
 
 
     <ImageView
